It is finally memorizing
Using crossentropy loss. Masking the input in the cross entropy loss
with -100 (perhaps so we can get a multidimensional loss for the entire
sequence? Not sure why this is better than chopping the tensor.

Using a smaller dimensionality of dmodel 64 and less decoder heads.
maybe the gradients were too small to propagate to the start of the
network?